+++ freeman-core/src/Core/Plugin.php
@@ -14,7 +14,7 @@ defined( 'ABSPATH' ) || exit;
  */
 final class Plugin {
 
-	const VERSION = '1.11.2';
+	const VERSION = '1.11.3';
 
 	/**
 	 * Singleton instance.